+++ lib/boolean_validation.dart
@@ -22,4 +22,6 @@ export 'src/validation_messages/message_replacements_keys.dart';
 export 'src/validation_messages/validation_messages.dart';
 export 'src/validation_messages/messages_provider.dart';
 
+export 'src/core/extensions/string_extension.dart';
+
 export 'src/core/common_libs.dart';